+++ rs/canister/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Include `identity_canister_id` when pushing magic links ([#19](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/19))
 - Move `EmailSenderConfig` into `api` package so that it is public ([#22](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/22))
 - Bump Rust version to 1.78.0 ([#25](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/25))
+- Transform HTTP response to only return the status code ([#26](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/26))
 
 ## [[0.4.0](https://github.com/open-chat-labs/ic-sign-in-with-email/releases/tag/v0.4.0)] - 2024-04-19